insert-extract-pgsql/requirements.txt
psycopg2-binary
python-dotenv
pep8
pylint
dotenv